move get_pem_format to utils, rename and make initialize as getter method
