+++ connectors/sources/tests/test_google_cloud_storage.py
@@ -32,7 +32,7 @@ def get_mocked_source_object():
         {"service_account_credentials": SERVICE_ACCOUNT_CREDENTIALS, "retry_count": 0}
     )
     mocked_gcs_object = GoogleCloudStorageDataSource(configuration=configuration)
-    mocked_gcs_object._initialize_configurations()
+    mocked_gcs_object.get_service_account_credentials()
     return mocked_gcs_object
 
 
@@ -73,9 +73,7 @@ async def test_empty_configuration():
 
 
 @pytest.mark.asyncio
-async def test_invalid_configuration():
-    """Tests that the service account credential is in invalid json format"""
-
+async def test_raise_on_invalid_configuration():
     # Setup
     configuration = DataSourceConfiguration(
         {"service_account_credentials": "{'abc':'bcd','cd'}"}
@@ -520,28 +518,3 @@ async def test_api_call_for_attribute_error(catch_stdout, patch_logger):
             userProject=mocked_gcs_object.user_project_id,
         ):
             print("Method called successfully....")
-
-
-def test_get_pem_format():
-    """This function tests prepare private key with dummy values"""
-    # Setup
-    expected_formated_pem_key = """-----BEGIN PRIVATE KEY-----
-PrivateKey
------END PRIVATE KEY-----"""
-    source = get_mocked_source_object()
-    private_key = "-----BEGIN PRIVATE KEY----- PrivateKey -----END PRIVATE KEY-----"
-
-    # Execute
-    formated_privat_key = source.get_pem_format(private_key, max_split=2)
-    assert formated_privat_key == expected_formated_pem_key
-
-    # Setup
-    expected_formated_certificate = """-----BEGIN CERTIFICATE-----
-Certificate1
-Certificate2
------END CERTIFICATE-----"""
-    private_key = "-----BEGIN CERTIFICATE----- Certificate1 Certificate2 -----END CERTIFICATE-----"
-
-    # Execute
-    formated_privat_key = source.get_pem_format(private_key, max_split=1)
-    assert formated_privat_key == expected_formated_certificate
